servers/aiohttp-polls/requirements.txt
psycopg2-binary
aiohttp
aiohttp_jinja2
aiopg[sa]
pyyaml

# ci
pytest
Faker

# devel
pylint
autopep8
